backend/requirements.txt
flask==3.1.3
flask-cors==6.0.2
pymongo==4.7.3
bcrypt==4.1.3
PyJWT==2.8.0
python-dotenv==1.0.1
